fix lint

+++ src/flatbuffer_types.rs
@@ -162,7 +162,7 @@ impl FlatBufferManifestCodec {
             next_wal_sst_id: manifest.wal_id_last_seen() + 1,
             last_compacted_wal_sst_id: manifest.wal_id_last_compacted(),
             last_clock_tick: manifest.last_clock_tick(),
-            last_seq: manifest.last_seq() as u64,
+            last_seq: manifest.last_seq(),
             checkpoints,
         };
         Manifest {
